Move EL tests to new JSPExpressionLanguageTests class
